requirements.txt
bs4
requests
html5lib
SQLAlchemy
psycopg2
selenium
webdriver-manager
python-dotenv